requirements.txt
asgiref==3.6.0
Django==4.1.7
django-environ==0.10.0
django-filter==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
